+++ src/torchmetrics/functional/classification/average_precision.py
@@ -98,7 +98,7 @@ def _average_precision_compute(
         if preds.ndim == target.ndim and target.ndim > 1:
             weights = target.sum(dim=0).float()
         else:
-            weights = _bincount(target, minlength=num_classes).float()
+            weights = _bincount(target, minlength=max(num_classes, 2)).float()
         weights = weights / torch.sum(weights)
     else:
         weights = None