+++ src/torchmetrics/utilities/data.py
@@ -253,15 +253,15 @@ def _bincount(x: Tensor, minlength: Optional[int] = None) -> Tensor:
     Returns:
         Number of occurrences for each unique element in x
     """
-    if x.is_cuda and deterministic():
-        if minlength is None:
-            minlength = len(torch.unique(x))
+    if minlength is None:
+        minlength = len(torch.unique(x))
+    if deterministic():
         output = torch.zeros(minlength, device=x.device, dtype=torch.long)
         for i in range(minlength):
             output[i] = (x == i).sum()
         return output
-    else:
-        return torch.bincount(x, minlength=minlength)
+    z = torch.zeros(minlength, device=x.device, dtype=x.dtype)
+    return z.index_add_(0, x, torch.ones_like(x))
 
 
 def allclose(tensor1: Tensor, tensor2: Tensor) -> bool:
@@ -269,3 +269,11 @@ def allclose(tensor1: Tensor, tensor2: Tensor) -> bool:
     if tensor1.dtype != tensor2.dtype:
         tensor2 = tensor2.to(dtype=tensor1.dtype)
     return torch.allclose(tensor1, tensor2)
+
+
+def _movedim(tensor: Tensor, dim1: int, dim2: int) -> tensor:
+    if _TORCH_GREATER_EQUAL_1_7:
+        return torch.movedim(tensor, dim1, dim2)
+    if dim2 >= 0:
+        dim2 += 1
+    return tensor.unsqueeze(dim2).transpose(dim2, dim1).squeeze(dim1)