+++ CHANGELOG.md
@@ -19,7 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--
+- Classification refactor (
+    [#1054](https://github.com/Lightning-AI/metrics/pull/1054),
+)
+
 
 -